refactor: rename soilId{Fragments,Service} to soilData{Fragments,Service}
