+++ src/soilId/soilDataFragments.ts
