+++ src/soilId/soilDataService.ts
